server/requirements.txt
pyzmq==19.0.2
tensorflow==1.14.0
sentencepiece
git+https://github.com/d2gex/pubsub-zmq.git@0.1.1#egg=pubsub-zmq